refactor: name list-types as the list-type owner
Move checked-state under features/list-types to match the spec.
Menu reads and converts nested list type through that owner instead
of calling outline list helpers itself.
