+++ src/client/editor/features/list-types/nested-list-type.ts
@@ -1,0 +1,10 @@
+import type { ListItemNode, ListType } from '@lexical/list';
+import { getNestedList } from '#client/editor/outline/selection/tree';
+
+export function $getNestedListType(contentItem: ListItemNode): ListType | null {
+  return getNestedList(contentItem)?.getListType() ?? null;
+}
+
+export function $setNestedListType(contentItem: ListItemNode, listType: ListType): void {
+  getNestedList(contentItem)?.setListType(listType);
+}
