+++ src/client/editor/runtime/nodes.ts
@@ -2,7 +2,7 @@ import type { InitialConfigType } from '@lexical/react/LexicalComposer';
 import { AutoLinkNode, LinkNode } from '@lexical/link';
 import { ListItemNode, ListNode } from '@lexical/list';
 
-import { ensureCheckedStateConfig } from '#client/editor/features/checklist/checked-state';
+import { ensureCheckedStateConfig } from '#client/editor/features/list-types/checked-state';
 import { ensureFoldStateConfig } from '#client/editor/outline/fold-state';
 import { DateNode } from '../features/date/date-node';
 import { BodyWrapperNode, NoteBodyNode } from '#client/editor/outline/note-body-node';
