+++ src/client/editor/shell/Editor.tsx
@@ -6,7 +6,7 @@ import { RichTextPlugin } from '@lexical/react/LexicalRichTextPlugin';
 import { useCallback, useState } from 'react';
 import { createEditorInitialConfig } from '#client/editor/runtime/config';
 import { CollaborationPlugin, useOfflineDocumentUnavailable } from '#client/editor/runtime/collaboration';
-import { CheckListPlugin } from '#client/editor/features/checklist/CheckListPlugin';
+import { CheckListPlugin } from '#client/editor/features/list-types/CheckListPlugin';
 import { IndentationPlugin } from '#client/editor/editing/indentation/IndentationPlugin';
 import { DevEditorSeam } from './DevEditorSeam';
 import { SelectionPlugin, SelectionInputPlugin } from '#client/editor/outline/selection/SelectionPlugin';
