+++ tests/unit/_support/lib/note.ts
@@ -17,7 +17,7 @@ import type { Outline } from '#tests-common/outline';
 import { extractOutlineFromEditorState } from '#tests-common/outline';
 import { findNearestListItem, getRootElementOrThrow } from './selection';
 import { $getNoteId } from '#client/editor/runtime/note-ids/note-id-state';
-import { $setNoteCheckedRaw } from '#client/editor/features/checklist/checked-state';
+import { $setNoteCheckedRaw } from '#client/editor/features/list-types/checked-state';
 import { expect } from 'vitest';
 import { COLLAPSE_STRUCTURAL_SELECTION_COMMAND } from '#client/editor/foundation/commands';
 export type { Outline, OutlineNode } from '#tests-common/outline';
